Replace InitOnceExecuteOnce with Meyer's Singleton
The InitOnceExecuteOnce function doesn't exist on Windows XP/Server 2003. It is
used to lazy initialise on first use. Whilst it is an elegant solution its
benefit in leveldb is negligable. I've replace it with a Meyer's Singleton that
is used to init the global at runtime rather than on first use. This loses the
lazy initialisation but gains the benefit of working on all flavours of
Windows.